refactored rough impl
