+++ src/main/java/com/laserfiche/api/client/oauth/TokenClientImpl.java
@@ -1,22 +1,13 @@
 package com.laserfiche.api.client.oauth;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.MapperFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.laserfiche.api.client.apiserver.ApiServer;
-import com.laserfiche.api.client.deserialization.OffsetDateTimeDeserializer;
-import com.laserfiche.api.client.deserialization.TokenClientObjectMapper;
-import com.laserfiche.api.client.httphandlers.HeadersImpl;
+import com.laserfiche.api.client.model.OauthClient;
 import com.laserfiche.api.client.model.AccessKey;
 import com.laserfiche.api.client.model.ApiException;
 import com.laserfiche.api.client.model.GetAccessTokenResponse;
 import com.laserfiche.api.client.model.ProblemDetails;
 import kong.unirest.Unirest;
 import kong.unirest.json.JSONObject;
-import org.threeten.bp.OffsetDateTime;
 
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
@@ -25,7 +16,7 @@ import static com.laserfiche.api.client.oauth.OAuthUtil.createBearer;
 import static com.laserfiche.api.client.oauth.OAuthUtil.getOAuthApiBaseUri;
 
 
-public class TokenClientImpl extends ApiServer implements TokenClient {
+public class TokenClientImpl extends OauthClient implements TokenClient {
     private String baseUrl;
 
     public TokenClientImpl(String regionalDomain) {
